+++ frontend/src/components/layout/header.tsx
@@ -1,15 +1,28 @@
 "use client";
-import { Settings, LogOut } from "lucide-react";
+import { Settings, LogOut, Copy, Check } from "lucide-react";
 import Link from "next/link";
 import { useSession, signIn, signOut } from "next-auth/react";
 import { Button } from "@/components/ui/button";
 
 import { ThemeToggle } from "@/components/layout/theme-changer";
 import { useProgress } from "@bprogress/next";
+import { ExtendedSession } from "@/lib/auth";
+import { useState } from "react";
 
 export function Header() {
-  const { data: session, status } = useSession();
+  const { data: session, status } = useSession() as {
+    data: ExtendedSession | null;
+    status: "authenticated" | "loading" | "unauthenticated";
+  };
   const { start } = useProgress();
+  const [isClicked, setIsClicked] = useState(false);
+  const accessToken = session?.accessToken;
+
+  const handleCopy = (accessToken: string) => {
+    navigator.clipboard.writeText(accessToken);
+    setIsClicked(true);
+    setTimeout(() => setIsClicked(false), 1500);
+  };
 
   return (
     <header className="flex h-16 items-center justify-between border-b-2 p-4 text-center">
@@ -20,6 +33,25 @@ export function Header() {
       </div>
       <div className="flex items-center gap-4">
         <ThemeToggle />
+        {accessToken && (
+          <Button
+            onClick={() => handleCopy(accessToken)}
+            className="flex w-[120px] items-center justify-center gap-2"
+            variant="outline"
+          >
+            {isClicked ? (
+              <>
+                <Check className="h-4 w-4" />
+                Copied !
+              </>
+            ) : (
+              <>
+                <Copy className="h-4 w-4" />
+                Copy Token
+              </>
+            )}
+          </Button>
+        )}
         {status === "loading" ? (
           <div className="h-5 w-5 animate-spin rounded-full border-2 border-gray-500 border-t-transparent" />
         ) : session ? (